Style: PerformanceList 스타일 분리
